# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/staker/delegation/check": {
            "get": {
                "description": "Check if a staker has an active delegation by the staker BTC address (Taproot only)\nOptionally, you can provide a timeframe to check if the delegation is active within the provided timeframe\nThe available timeframe is \"today\" which checks after UTC 12AM of the current day",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Staker BTC address in Taproot format",
                        "name": "address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "today"
                        ],
                        "type": "string",
                        "description": "Check if the delegation is active within the provided timeframe",
                        "name": "timeframe",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/handlers.Result"
                        }
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_babylonchain_xchains-api_internal_types.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_babylonchain_xchains-api_internal_types.Error": {
            "type": "object",
            "properties": {
                "err": {},
                "errorCode": {
                    "$ref": "#/definitions/types.ErrorCode"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.Result": {
            "type": "object",
            "properties": {
                "data": {},
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "types.ErrorCode": {
            "type": "string",
            "enum": [
                "INTERNAL_SERVICE_ERROR",
                "VALIDATION_ERROR",
                "NOT_FOUND",
                "BAD_REQUEST",
                "FORBIDDEN"
            ],
            "x-enum-varnames": [
                "InternalServiceError",
                "ValidationError",
                "NotFound",
                "BadRequest",
                "Forbidden"
            ]
        }
    }
}